==> 0_Cleaning_Data.ipynb <==
# Data Preperation & Cleaning

## Import modules

import pandas as pd

## Read data from csv file

df_txt = pd.read_csv("data/data.tsv.txt", sep="<br>", header=None, names=['0', '1', '2'])

df_txt.sample(5)

df = pd.read_csv("data/haikus.csv")

df.head()

## Get only haikus with the syllable structure 5-7-5

df_575_only = df[(df['0_syllables'] == '5') & (df['1_syllables'] == '7') & (df['2_syllables'] == '5')]

## Drop not needed columns

pd.options.mode.chained_assignment = None # supress warning
df_575_only.drop(columns=["source", '0_syllables', '1_syllables', '2_syllables'], inplace=True)

# Merge dfs

df = df_575_only.append(df_txt)

## Make every value in columns to lowercase

df = df.applymap(lambda s: s.lower() if type(s) == str else s)

## Remove unnecessary chars

df['0'] = df['0'].str.replace('[^\w\s]', '')
df['1'] = df['1'].str.replace('[^\w\s]', '')
df['2'] = df['2'].str.replace('[^\w\s]', '')

df = df[~df['0'].str.contains('_')]
df = df[~df['1'].str.contains('_')]
df = df[~df['2'].str.contains('_')]

# Remove rows with numbers

df = df[~df['0'].str.contains(r'\d')]
df = df[~df['1'].str.contains(r'\d')]
df = df[~df['2'].str.contains(r'\d')]

## Drop index and save to csv

df.to_csv('data/out.csv', index=False)

df.shape
