==> 5_w2v_Tensorboard.ipynb <==
# Model training

## Import packages

import tensorflow as tf
import pandas as pd
import os
from tensorflow import keras
from tensorflow.keras import layers
import numpy as np

import random
import re
import matplotlib.pyplot as plt
from scipy import sparse
from tqdm import tqdm
from scipy import sparse

# Deep learning: 
from keras.models import Input, Model
from keras.layers import Dense

#Activate tensorboard extension
try:
  # %tensorflow_version only exists in Colab.
  %tensorflow_version 2.x
except Exception:
  pass

%load_ext tensorboard

#Import necessary libs
import tensorflow_datasets as tfds
from tensorboard.plugins import projector

print("Num GPUs Available: ", len(tf.config.experimental.list_physical_devices('GPU')))

## Read data

# Datensatz einlesen
df = pd.read_csv('data/out.csv')

## Transform dataframe

# Dataframe mit 3 Spalten. Werden so gejoint, dass ein neues Dataframe mit ein Haiku pro Zeile erstellt wird
df = df[['0', '1', '2']].agg(lambda x: ' '.join(x.values), axis=1)
# Dataframe to list [[]] -> []
haikus = df.values.tolist()

# ---------------------------------------------------
# Erst mal nur mit 1000 Haikus!
haikus = haikus[:5000]
# ---------------------------------------------------

# alle Haikus in Array
print(haikus[-10:])
number_of_haikus = len(haikus)
print('number of haikus: ' + str(number_of_haikus))

## Clean text 

huge_string = ' SATZENDE SATZANFANG '.join(haikus)
huge_string_cleaned = re.sub('[.,_]', '', huge_string)

# TODO remove stopwords!

huge_list = huge_string_cleaned.split(' ')

#Get vocabulary
vocab = sorted(set(huge_list))
print(vocab[:20])
vocab_size = len(vocab)
print(vocab_size)

vocab = np.array(vocab)

## Create context windows

# Defining the window for context
window = 3

word_lists = []

# Creating a context dictionary
for i, word in enumerate(huge_list):
    for w in range(window):
        # Getting the context that is ahead by *window* words
        if i + 1 + w < len(huge_list): 
            word_lists.append([word] + [huge_list[(i + 1 + w)]])
        # Getting the context that is behind by *window* words    
        if i - w - 1 >= 0:
            word_lists.append([word] + [huge_list[(i - w - 1)]])

print(word_lists[-100:])

## One-hot encode

# Creating the dictionary for the unique words
unique_word_dict = {}
for i, word in enumerate(vocab):
    unique_word_dict.update({
        word: i
    })
    
print(unique_word_dict.get('a'))

# Defining the number of features (unique words)
n_words = len(unique_word_dict)

# Getting all the unique words 
words = list(unique_word_dict.keys())

# word_lists1 = word_lists[:int(len(word_lists)/10000)]
print('len(word_lists): ' + str(len(word_lists)))
#print('len(word_lists1): ' + str(len(word_lists1)))


# Creating the X and Y matrices using one hot encoding
X = np.zeros((len(word_lists), n_words), dtype=bool)
Y = np.zeros((len(word_lists), n_words), dtype=bool)

for i, word_list in tqdm(enumerate(word_lists)):
    # Getting the indices
    main_word_index = unique_word_dict.get(word_list[0])      #First word in tupel
    context_word_index = unique_word_dict.get(word_list[1])   #Second word in Tupel

    # One hot encoding the main word
    X[i, main_word_index] = 1

    # One hot encoding the Y matrix words 
    Y[i, context_word_index] = 1


#! Hier schmiert der Kernel bei der Iteration immer ab.. 

train_dataset = tf.data.Dataset.from_tensor_slices((X, Y))

train_dataset

## The model

# Defining the size of the embedding (Often in practice, pre-trained word embeddings are used with typical word embedding dimensions being either 100, 200 or 300.)
embed_size = 100

# Defining the neural network
inp = Input(shape=(X.shape[1],))
x = Dense(units=embed_size, activation='linear')(inp)
x = Dense(units=Y.shape[1], activation='softmax')(x)
model = Model(inputs=inp, outputs=x)

model.compile(loss = 'categorical_crossentropy', optimizer = 'adam')

train_in_batches():
    

# Optimizing the network weights
model.fit(
    train_dataset,
    batch_size=256,
    epochs=50
    )

model.summary()

# Obtaining the weights from the neural network. 
# These are the so called word embeddings

# The input layer 
weights = model.get_weights()[0]

# Creating a dictionary to store the embeddings in. The key is a unique word and 
# the value is the numeric vector
embedding_dict = {}
for word in words: 
    embedding_dict.update({
        word: np.asarray(weights[unique_word_dict.get(word)])
        })

import matplotlib.pyplot as plt

plt.figure(figsize=(10, 10))

for word in list(unique_word_dict.keys()):
    coord = embedding_dict.get(word)
    plt.scatter(coord[0], coord[1])
    plt.annotate(word, (coord[0], coord[1]))

# Save metadata into tsv file
pd.DataFrame(embedding_dict.keys()).to_csv("model_dir/metadata.tsv", sep = '\t', index=False)

# Save embeddings_vectors into tsv file
pd.DataFrame(embedding_dict.values()).to_csv("model_dir/vectors_2.tsv", sep = '\t', index=False)
