==> 6_Haikus_with_w2v_Wordembeddings.ipynb <==
# Model training

## Import packages

!pip install --upgrade gensim

import tensorflow as tf
import pandas as pd
import os
from tensorflow import keras
from tensorflow.keras import layers
import numpy as np
import random
import re
import matplotlib.pyplot as plt
from scipy import sparse
from tqdm import tqdm
import gensim
from keras.callbacks import LambdaCallback

# Deep learning: 
from keras.models import Input, Model
from keras.layers import Dense

## Read data

# Datensatz einlesen
df = pd.read_csv('data/out.csv')

## Transform dataframe

# Dataframe mit 3 Spalten. Werden so gejoint, dass ein neues Dataframe mit ein Haiku pro Zeile erstellt wird
df = df[['0', '1', '2']].agg(lambda x: ' \n '.join(x.values), axis=1)
# Dataframe to list [[]] -> []
haikus = df.values.tolist()

# alle Haikus in Array
print(haikus[:10])
#number_of_haikus = len(haikus)
number_of_haikus = 400000

haikus = haikus[:number_of_haikus]
print('number of haikus: ' + str(number_of_haikus))

# die Haikus cleanen und selber auch noch mal als Wort-Arrays in großen Array
haikus = np.array(haikus)

def clean_and_split(sentence):
    result = list(filter(''.__ne__, re.sub('[.,_]', '', sentence).split(' ')))
    result.append(';')
    return result

haikus = list(map(lambda x: clean_and_split(x), haikus))

print(haikus[:10])

## The word_model

# train word_model
word_model = gensim.models.Word2Vec(haikus, min_count=1) # min count automatisch groesser, discarded alle woerter die weniger vorkommen

# summarize the loaded word_model
print(word_model)

# access vector for one word
print('vector for \'girl\':')
print(word_model.wv['girl'])

print('top 10 words most similar to \'girl\':')
word_model.wv.most_similar('girl', topn=10)

# similarity between two words
print('similarity between \'go\' and \'walk\' (regarding the haikus):')
print(word_model.wv.similarity(w1='go', w2='walk'))
print()

print('similarity between \'go\' and \'laugh\' (regarding the haikus):')
print(word_model.wv.similarity(w1='go', w2='laugh'))
print()

print('similarity between \'go\' and \'go\':')
print(word_model.wv.similarity(w1='go', w2='go'))

# save word_model
#word_model.save('w2v_model.bin')

# load model
#new_model = Word2Vec.load('w2v_model.bin')
#print(new_model)

# extract the words & their vectors, as numpy arrays
vectors = np.asarray(word_model.wv.vectors)
labels = np.asarray(word_model.wv.index_to_key)  # fixed-width numpy strings

print('vectors:')
print(vectors[:2])
print()
print('labels:')
print(labels[:10])

len(vectors)

len(labels)

# https://projector.tensorflow.org/

# Save metadata (labels) into tsv file
pd.DataFrame(labels).to_csv("model_dir/metadata.tsv", sep = '\t', index=False)

# Save vectors into tsv file
pd.DataFrame(vectors).to_csv("model_dir/vectors.tsv", sep = '\t', index=False)

## Creating Model for HaikuGen

# Maximale Anzahl der Wörter in einem Haiku aus Datenset
max_haiku_len = len(max(haikus, key=len))
max_features = 20000

haikus[4]

word_model.wv["out"]

def word2idx(word):
    return word_model.wv.key_to_index[word]
def idx2word(idx):
    return word_model.wv.index_to_key[idx]

#haikus_combined =  sum(haikus, [])

haikus_combined = [item for sublist in haikus for item in sublist]
# cut the text in semi-redundant sequences of seq_len characters
print(haikus_combined[1:20])
seq_len = max_haiku_len
step = 7
# Input String
sequences = []
#Output character
next_words = []
for i in range(0, len(haikus_combined) - seq_len, step):
    sequences.append(haikus_combined[i : i + seq_len])
    next_words.append(haikus_combined[i + seq_len])
print("Number of sequences:", len(sequences))

train_x = np.zeros([len(sequences), max_haiku_len], dtype=np.int32)
train_y = np.zeros([len(next_words)], dtype=np.int32)




for i, sequence in enumerate(sequences):
    for t, word in enumerate(sequence):
         train_x[i, t] = word2idx(word)
    train_y[i] = word2idx(next_words[i])
print('train_x shape:', train_x.shape)
print('train_y shape:', train_y.shape)


#print('\nPreparing the data for LSTM...')




#train_x = np.zeros([len(haikus), max_haiku_len], dtype=np.int32)
#train_y = np.zeros([len(haikus)], dtype=np.int32)
#for i, haiku in enumerate(haikus):
#    for t, word in enumerate(haiku[:-1]):
#        train_x[i, t] = word2idx(word)
#    train_y[i] = word2idx(haiku[-1])
#print('train_x shape:', train_x.shape)
#print('train_y shape:', train_y.shape)

pretrained_weights = word_model.wv.vectors
vocab_size = pretrained_weights.shape[0]
emdedding_size = pretrained_weights.shape[1]
print('Result embedding shape:', pretrained_weights.shape)

from keras.models import Sequential
from keras.layers import Dense, Embedding, Activation 
from keras.layers import LSTM

print('Build model...')
model = Sequential()
model.add(Embedding(input_dim=vocab_size, output_dim=emdedding_size, weights=[pretrained_weights]))
model.add(LSTM(emdedding_size, dropout=0.2, recurrent_dropout=0.2,return_sequences=True))
model.add(LSTM(emdedding_size*2, dropout=0.2, recurrent_dropout=0.2))
model.add(Dense(vocab_size)),
model.add(Activation('softmax')),
model.compile(optimizer='adam', loss='sparse_categorical_crossentropy')

#print('\nTraining LSTM...')
#model = Sequential()
#model.add(Embedding(input_dim=vocab_size, output_dim=emdedding_size, weights=[pretrained_weights]))
#model.add(LSTM(units=emdedding_size))
#model.add(Dense(units=vocab_size))
#model.add(Activation('softmax'))

model.summary()

#model.fit(train_x, train_y,
#          batch_size=512,
#          epochs=1)

def sample(preds, temperature=1.0):
    if temperature <= 0:
        return np.argmax(preds)
    preds = np.asarray(preds).astype('float64')
    preds = np.log(preds) / temperature
    exp_preds = np.exp(preds)
    preds = exp_preds / np.sum(exp_preds)
    probas = np.random.multinomial(1, preds, 1)
    return np.argmax(probas)

def generate_next(text, num_generated=10):
    word_idxs = [word2idx(word) for word in text.lower().split()]
    for i in range(num_generated):
        prediction = model.predict(x=np.array(word_idxs))
        idx = sample(prediction[-1], temperature=0.7)
        word_idxs.append(idx)
    return ' '.join(idx2word(idx) for idx in word_idxs)

def generate_haikus(generate_words, temperature=1.0):

    start_index = random.randint(0, train_x.shape[0] - seq_len - 1)
    generated = ""

    seed =  train_x[start_index:start_index+1]

    for i in range(generate_words):
        preds = model.predict(seed)
        pred_index = sample(preds[-1])
        pred_word = idx2word(pred_index)
    
    
        seed_tmp = np.concatenate((seed[0][1:], [pred_index]))
        seed[0] = seed_tmp
    
        generated += pred_word + " "
    
        if pred_word == ";":
            generated += "\n----------------------------------------\n"

    print(generated)
    
    


epochs = 20
batch_size = 512

for epoch in range(epochs):
    print()
    print()
    print(f"EPOCH:{epoch}")
    model.fit(train_x, train_y, 
              batch_size=batch_size, 
              epochs=1)
    
    model.save('myModelEmbedding.h5')

    generate_haikus(30)


generate_haikus(200)
