==> 8_Visualize.ipynb <==
!pip install --upgrade gensim

!pip install -q gpt-2-simple

import tensorflow as tf
import pandas as pd
import os
from tensorflow import keras
from tensorflow.keras import layers
import numpy as np
import random
import re
import ipywidgets as widgets
import gensim
import time
import gpt_2_simple as gpt2
from datetime import datetime

# Datensatz einlesen
df = pd.read_csv('data/out.csv')

# Dataframe mit 3 Spalten. Werden so gejoint, dass ein neues Dataframe mit ein Haiku pro Zeile erstellt wird
df = df[['0', '1', '2']].agg(lambda x: '\n'.join(x.values), axis=1)
# Dataframe to list [[]] -> []
array_of_poems = df.values.tolist()

# num_poems an Haikus mit Semicolon zwshceneinander zusammenfuegen
num_poems = 10000

text = ';'.join(array_of_poems[:num_poems])

# Anzahl der unterschielichen characters im gesamt datensatz herausfinden
vocab = sorted(set(text))
vocab_size = len(vocab)
print(f'{vocab_size} unique chars')
print(vocab)

# Dictionary erstellen. Jeder character wird mit einer Zahl nummeriert
char_indices = {c: i for i, c in enumerate(vocab)}
indices_char = {i: c for i, c in enumerate(vocab)}

# cut the text in semi-redundant sequences of seq_len characters
seq_len = 100

model = keras.models.load_model('myModelTuner.h5')

model2 = keras.models.load_model('myModelEmbedding.h5')

# Standartfunktion Probability array to onehot to integerencoded char
# [0.3, 0.2, 0.1, 0.4] -> [0, 0, 0, 1] -> return 4 (stelle, an der 1)
def sample(prob, temperature=1.0):
    np.seterr(divide = 'ignore')
    #np.seterr(divide = 'warn') 
    # helper function to sample an index from a probability array
    prob = np.asarray(prob).astype("float64")
    prob = np.log(prob) / temperature
    exp_prob = np.exp(prob)
    prob = exp_prob / np.sum(exp_prob)
    probas = np.random.multinomial(1, prob)
    return np.argmax(probas)


output = widgets.Output()
@output.capture(clear_output=True)
def generate_haiku_characterbased(haiku_start, temperature=1.0):
    lenString = 300
    randomNum = random.randint(0, len(text) - lenString)
    findTemplate = text[randomNum: randomNum + lenString]

    pattern = '(.{99};)'
    seed = re.search(pattern, findTemplate, re.DOTALL).group()

    poem = format(haiku_start.value)
    seed = seed[len(poem):] + poem


    print(poem, end='')
    
    while(len(poem) == 0 or poem[-1] != ';'):
        x_pred = np.zeros((1, len(seed), vocab_size))
        for t, char in enumerate(seed):
            x_pred[0, t, char_indices[char]] = 1
        preds = model.predict(x_pred, verbose=0)[0]
        next_index = sample(preds, temperature)
        next_char = indices_char[next_index]
        seed = seed[1:] + next_char
        poem += next_char
        time.sleep(200/1000)
        if(next_char != ';'):
            print(next_char, end='')



df2 = pd.read_csv('data/out.csv')

# Dataframe mit 3 Spalten. Werden so gejoint, dass ein neues Dataframe mit ein Haiku pro Zeile erstellt wird
df2 = df2[['0', '1', '2']].agg(lambda x: ' \n '.join(x.values), axis=1)
# Dataframe to list [[]] -> []
haikus = df2.values.tolist()

# alle Haikus in Array
#number_of_haikus = len(haikus)
number_of_haikus = 10000

haikus = haikus[:number_of_haikus]


print('number of haikus: ' + str(number_of_haikus))

# die Haikus cleanen und selber auch noch mal als Wort-Arrays in großen Array
haikus = np.array(haikus)

def clean_and_split(sentence):
    result = list(filter(''.__ne__, re.sub('[.,_]', '', sentence).split(' ')))
    result.append(';')
    return result

haikus = list(map(lambda x: clean_and_split(x), haikus))

word_model = gensim.models.Word2Vec.load('w2v_model.bin')

def word2idx(word):
    return word_model.wv.key_to_index[word]
def idx2word(idx):
    return word_model.wv.index_to_key[idx]

#haikus_combined =  sum(haikus, [])

haikus_combined = [item for sublist in haikus for item in sublist]
# cut the text in semi-redundant sequences of seq_len characters
seq_len = 20
step = 1
# Input String
sequences = []

for i in range(0, len(haikus_combined) - seq_len, step):
    sequences.append(haikus_combined[i : i + seq_len])

train_x = np.zeros([len(sequences), seq_len], dtype=np.int32)



for i, sequence in enumerate(sequences):
    for t, word in enumerate(sequence):
         train_x[i, t] = word2idx(word)



generation = []
for i, sequence in enumerate(train_x):
    if(train_x[i, seq_len- 1] == word2idx(";")):
        generation.append(sequence)
        
print(len(generation))

def sample2(preds, temperature=1.0):
    np.seterr(divide = 'ignore')
    #np.seterr(divide = 'warn') 
    if temperature <= 0:
        return np.argmax(preds)
    preds = np.asarray(preds).astype('float64')
    preds = np.log(preds) / temperature
    exp_preds = np.exp(preds)
    preds = exp_preds / np.sum(exp_preds)
    probas = np.random.multinomial(1, preds, 1)
    return np.argmax(probas)

def generate_next(text, num_generated=10):
    word_idxs = [word2idx(word) for word in text.lower().split()]
    for i in range(num_generated):
        prediction = model.predict(x=np.array(word_idxs))
        idx = sample(prediction[-1], temperature=0.7)
        word_idxs.append(idx)
    return ' '.join(idx2word(idx) for idx in word_idxs)

output2 = widgets.Output()
@output2.capture(clear_output=True)
def generate_haiku_wordbased(haiku_start, temperature=1.0):
    
    haiku_start = format(haiku_start.value)
    
    split_words = list(filter(''.__ne__, re.sub('[.,_]', '', haiku_start).split(' ')))
    
    start_index = random.randint(0, len(generation)-1)
    generate_with =  generation[start_index]
    
    for word in split_words:
        generate_with = np.append(generate_with[1:], word2idx(word))


    seed = np.zeros([1, seq_len], dtype=np.int32)
    for i in range(seq_len):
        seed[0][i] = generate_with[i]

    pred_word = ""
    print(haiku_start, end=' ')
    while(pred_word != ";"):
        preds = model2.predict(seed)
        pred_index = sample2(preds[-1])
        pred_word = idx2word(pred_index)
        seed_tmp = np.concatenate((seed[0][1:], [pred_index]))
        seed[0] = seed_tmp
        time.sleep(500/1000)
        if(pred_word != ";"):
            print(pred_word, end=' ')
    
        

#sess = gpt2.start_tf_sess()
sess = gpt2.reset_session(sess)
gpt2.load_gpt2(sess, run_name='run1')

output3 = widgets.Output()
@output3.capture(clear_output=True)
def generate_haiku_transformer(haiku_start, temperature=1.0):
    haiku_start_format = format(haiku_start.value)
    
    lenString = 300
    randomNum = random.randint(0, len(text) - lenString)
    findTemplate = text[randomNum: randomNum + lenString]
    
    pattern = ';([^;]*;)'
    seed = re.search(pattern, findTemplate, re.DOTALL).group(1)
    
    generated_transformer = gpt2.generate(sess,
              length=30,
              temperature=1.0,
              prefix=seed + "\n\n" + haiku_start_format,
              nsamples=1,
              batch_size=1,
              return_as_list=True)[0]
    pattern_transformer = ';\n\n([^;]*);'
    haiku_transformer = re.search(pattern_transformer, generated_transformer, re.DOTALL).group(1)
    print(haiku_transformer)
    

input_haiku = widgets.Text(value="",
                           placeholder='Type in the start of your Haiku...',
                           disable=False
)

description = widgets.HTML('Generate a characterbased Haiku:')

display(description)
display(input_haiku)

input_haiku.on_submit(generate_haiku_characterbased)
display(output)

input_haiku = widgets.Text(value="",
                           placeholder='Type in the start of your Haiku...',
                           disable=False
)

description = widgets.HTML('Generate a word-based Haiku:')

display(description)
display(input_haiku)

input_haiku.on_submit(generate_haiku_wordbased)
display(output2)

input_haiku = widgets.Text(value="",
                           placeholder='Type in the start of your Haiku...',
                           disable=False
)

description = widgets.HTML('Generate Transformer-Haiku:')

display(description)
display(input_haiku)

input_haiku.on_submit(generate_haiku_transformer)
display(output3)
